fix: Override attribute type during discovery
Attributes that are array-like may get falsly interpreted as scalar if they appear as a singleton in a record.
Previously, attribute discovery assigned the type of the first instance observed. Now, we override an attr
already inferred as String with Array once an array instance is discovered. This will improve the
accuracy of attribute type inference but is not foolproof: users can still manually upcast to Array if the
sniffer fails to discover if a given attribute should be of type Array.

Also, previously, this bug was silent in the case of GTF: if an attr were falsly defined as String but
actually occured with multiplicity greater than 1 in some records, the corresponding value was getting
overriden during batch creation rather than causing an error.