chore: Migrate noodles 0.102 (#150)
* chore: Migrate alignment, variant, bed, sequence to noodles 0.102

* Migrate gtf/gff to noodles 0.102

* fix: Override attribute type during discovery

Attributes that are array-like may get falsely interpreted as scalar if they appear as a singleton in a record. Previously, attribute discovery assigned the type of the first instance observed. Now, we override an attr already inferred as String with Array once an array instance is discovered. This will improve the accuracy of attribute type inference but is not foolproof: users can still manually upcast to Array if the sniffer fails to discover if a given attribute should be of type Array.

Also, previously, this bug was silent in the case of GTF during record batch creation: if an attr were falsly defined as String but actually occured with multiplicity greater than 1 in some records, only the first value was recorded. Now, an error is raised.

* Update py-oxbow to support noodles migration

* Update r-oxbow to support noodles migration